Identity and operators in separate files
